requirements.txt
openai
PyMuPDF
faiss-cpu
sentence-transformers
prompt_toolkit
python-dotenv